command-server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
docopt==0.6.2
fastapi==0.110.0
h11==0.14.0
idna==3.6
pydantic==2.6.3
pydantic_core==2.16.3
sniffio==1.3.1
starlette==0.36.3
stomp.py==8.1.0
typing_extensions==4.10.0
uvicorn==0.28.0
websocket-client==1.7.0
